Added new Ribber SCAD files
This is a semi-comprehensive reworking of the needlebed system, seeking to alter how the modules are connected to each other for a stronger and more robust connection, to be more suitable for double-bed operation.

Additionally, an STL file for an angled mounting bracket is included. This was modeled in FreeCAD rather than OpenSCAD; trying to get those compound angles working properly was an absolute nightmare in SCAD, and I'm more familiar with FreeCAD. The FreeCAD design file is included, in the event you wish to change how the bracket mounts to other things, or the angle between the beds, or the amount of needle overlap.

None of this has been printed yet. It's printing right now, so I will test it and make tweaks as necessary in a future update.